Proof of concept: an include feature for justfiles.
A fast proof that it's straightforward to do a simple
include feature for Just. This design takes all lines that
look like this: `#include "relative/path/to/file"`

and replaces them with the contents of the pointed-to file.
The resulting buffer is passed along to be parsed and executed
as before. That's it: no attempt to look at the file or do anything
with the contents. This is pure concatenation.